+++ zebra-rpc/src/methods.rs
@@ -2713,6 +2713,25 @@ where
             Some(&coinbase_cache),
         );
 
+        #[cfg(test)]
+        let (dependency_depths, selected_txs): (Vec<_>, Vec<_>) = mempool_txs.into_iter().unzip();
+        #[cfg(not(test))]
+        let selected_txs = mempool_txs;
+
+        let selected_txs = types::get_block_template::tachyon::aggregate_transactions(
+            self.network.clone(),
+            height,
+            chain_info.tip_hash,
+            read_state,
+            selected_txs,
+        )
+        .await;
+
+        #[cfg(test)]
+        let mempool_txs: Vec<_> = dependency_depths.into_iter().zip(selected_txs).collect();
+        #[cfg(not(test))]
+        let mempool_txs = selected_txs;
+
         tracing::debug!(
             selected_mempool_tx_hashes = ?mempool_txs
                 .iter()